prototype offsets as symbols
